feat: Integrated Yan's advice toolkit
The search knowledge base functionality isn't working properly. It's returning a search result that does not fit in the token window.
This is a chunking and retrieval issue we shall fix in the future
As for the search functionality, it's unclear how we should use it at this stage. Ideally, we need to use it for less and less stuff